Revert to torch_dtype for transformers 4.49 compatibility

+++ training/train_hf.py
@@ -193,7 +193,7 @@ def train_bridge_phase(ujamaa_model, bridge: AnthosProjectionBridge, dataset, ph
         if step >= phase_cfg["steps"]:
             break
         # Dummy forward for now — replace with real paired Anthos embeddings
-        fake_hidden = torch.randn(1, 32, 2560, dtype=dtype)
+        fake_hidden = torch.randn(1, 32, 2560, torch_dtype=dtype)
         perception = bridge(fake_hidden)
         # Self-supervised: perception tokens should be spread (not collapsed)
         loss = -perception.var(dim=1).mean()
@@ -232,15 +232,15 @@ def main():
     processor = AutoProcessor.from_pretrained(model_id, trust_remote_code=True)
     try:
         from transformers import Qwen2_5_VLForConditionalGeneration
-        model = Qwen2_5_VLForConditionalGeneration.from_pretrained(model_id, dtype=dtype, trust_remote_code=True)
+        model = Qwen2_5_VLForConditionalGeneration.from_pretrained(model_id, torch_dtype=dtype, trust_remote_code=True)
     except (ImportError, Exception):
         try:
             if AutoModelForVision2Seq is not None:
-                model = AutoModelForVision2Seq.from_pretrained(model_id, dtype=dtype, trust_remote_code=True)
+                model = AutoModelForVision2Seq.from_pretrained(model_id, torch_dtype=dtype, trust_remote_code=True)
             else:
                 raise ImportError
         except Exception:
-            model = AutoModelForCausalLM.from_pretrained(model_id, dtype=dtype, trust_remote_code=True)
+            model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=dtype, trust_remote_code=True)
     model = apply_lora(model, cfg)
     model.print_trainable_parameters()
